transforma helper functions em middlewares req 04
