--- helpers/index.js
@@ -1,60 +1,0 @@
-const checkToken = (token, response) => {
-  if (!token) return response.status(401).json({ message: 'Token não encontrado' });
-  if (typeof token !== 'string' || token.length !== 16) {
-    return response.status(401).json({ message: 'Token inválido' });
-  }
-};
-
-const checkName = (name, response) => {
-  if (!name) return response.status(400).json({ message: 'O campo "name" é obrigatório' });
-  if (name.lenth < 3) {
-    return response.status(400).json({ message: 'O "name" deve ter pelo menos 3 caracteres' });
-  }
-};
-
-const checkTalk = (talk, response) => {
-  if (!talk || talk === {} || !talk.watchedAt || !talk.rate) {
-    return response
-      .status(400)
-      .json({
-        message: 'O campo "talk" é obrigatório e "watchedAt" e "rate" não podem ser vazios',
-        });
-  }
-};
-
-const checkAge = (age, response) => {
-  if (!age || age === '') {
-    return response.status(400).json({ message: 'O campo "age" é obrigatório' });
-  }
-  if (age < 18) {
-    return response.status(400).json({ message: 'A pessoa palestrante deve ser maior de idade' });
-  }
-};
-
-const checkTalWatchedAt = (watchedAt, response) => {
-  const dateRegex = /^(0?[1-9]|[12][0-9]|3[01])[/-](0?[1-9]|1[012])[/-]\d{4}$/;
-  const isValidDate = dateRegex.test(watchedAt);
-  if (!isValidDate) {
-    return response.status(400).json({
-    message: 'O campo "watchedAt" deve ter o formato "dd/mm/aaaa"',
-    });
-  }
-};
-
-const checkTalkRate = (rate, response) => {
-  if (typeof rate !== 'number' || rate < 1 || rate > 5) {
-    return response.status(400)
-      .json({
-        message: 'O campo "rate" deve ser um inteiro de 1 à 5',
-      });
-  }
-};
-
-module.exports = {
-  checkToken,
-  checkName,
-  checkAge,
-  checkTalk,
-  checkTalWatchedAt,
-  checkTalkRate,
-};
