+++ middlewares/checkNameMiddleware.js
@@ -1,0 +1,8 @@
+module.exports = (request, response, next) => {
+  const { name } = request.body;
+  if (!name) return response.status(400).json({ message: 'O campo "name" é obrigatório' });
+  if (name.lenth < 3) {
+    return response.status(400).json({ message: 'O "name" deve ter pelo menos 3 caracteres' });
+  }
+  next();
+};
